+++ function/python/brightics/function/classification/xgb_classification.py
@@ -2,7 +2,7 @@ import pandas as pd
 import matplotlib.pyplot as plt
 from xgboost import XGBClassifier
 from xgboost import plot_importance, plot_tree
-from brightics.common.report import ReportBuilder, strip_margin, pandasDF2MD, plt2MD, dict2MD
+from brightics.common.repr import BrtcReprBuilder, strip_margin, pandasDF2MD, plt2MD, dict2MD
 from brightics.function.utils import _model_dict
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
@@ -68,7 +68,7 @@ def _xgb_classification_train(table, feature_cols, label_col, max_depth=3, learn
 #     get_param_df = pd.DataFrame(data=get_param_list, columns=['parameter', 'value'])
     feature_importance_df = pd.DataFrame(data=feature_importance, index=feature_cols).T
     
-    rb = ReportBuilder()
+    rb = BrtcReprBuilder()
     rb.addMD(strip_margin("""
     | ## XGB Classification Train Result
     |
@@ -85,7 +85,7 @@ def _xgb_classification_train(table, feature_cols, label_col, max_depth=3, learn
                table_feature_importance=pandasDF2MD(feature_importance_df, 20),
                list_parameters=params            
                )))     
-    model['report'] = rb.get()   
+    model['_repr_brtc_'] = rb.get()   
                
     return {'model' : model}
