+++ function/python/brightics/function/statistics/anova.py
@@ -1,4 +1,4 @@
-from brightics.common.report import ReportBuilder, strip_margin, plt2MD, \
+from brightics.common.repr import BrtcReprBuilder, strip_margin, plt2MD, \
     pandasDF2MD, keyValues2MD
 
 from scipy.stats import bartlett
@@ -36,7 +36,7 @@ def bartletts_test(table, response_cols, factor_col):
     result = dict()
     result['result_table'] = result_table
     
-    rb = ReportBuilder()
+    rb = BrtcReprBuilder()
     rb.addMD(strip_margin("""
     ## Bartlett's Test Result
     | - H0: k population variances are equal.
@@ -45,13 +45,13 @@ def bartletts_test(table, response_cols, factor_col):
     | {result_table}
     """.format(result_table=pandasDF2MD(result_table))))
     
-    result['report'] = rb.get()
+    result['_repr_brtc_'] = rb.get()
         
     return {'result': result}
     
     
 def oneway_anova(table, response_cols, factor_col):
-    rb = ReportBuilder()
+    rb = BrtcReprBuilder()
     rb.addMD(strip_margin("""
     ## One-way Analysis of Variance Result
     """))
@@ -107,12 +107,12 @@ def oneway_anova(table, response_cols, factor_col):
         
         result['_grouped_data'][response_col]['p_value'] = p_value
         
-    result['report'] = rb.get()
+    result['_repr_brtc_'] = rb.get()
     return {'result': result}
     
     
 def tukeys_range_test(table, response_cols, factor_col, alpha=0.05):
-    rb = ReportBuilder()
+    rb = BrtcReprBuilder()
     rb.addMD("""## Tukey's range test Result""")
     
     for response_col in response_cols:
@@ -126,4 +126,4 @@ def tukeys_range_test(table, response_cols, factor_col, alpha=0.05):
         rb.addPlt(plt)
         plt.clf()
     
-    return {'result': {'report': rb.get()}}
+    return {'result': {'_repr_brtc_': rb.get()}}
